modify

+++ shell/merge.sh
@@ -7,21 +7,20 @@ function usage()
 {
     cat <<EOF
 Usage: `basename $0` [-f configfile]
-    The default configfile is merge.conf
+    Merge files and modify some char, the default configfile is merge.conf
     FILE FORMAT:
-        new_file    file_1:offset_1,file_2:offset_2[,...]     [new_char1:offset_char1[,...]]
+        new_file;   file_1:offset_1, file_2:offset_2[ ,...];    [new_char1:offset_char1[ ,...]]
         For example:
-        A.bin  a1.bin:0,a2.bin:0x8000,a3.bin:0x10000,a4.bin:0x18000 
-        B.bin  b1.bin:0,b2.bin:0x8000                                 \x0d:0x1234,\xfa:0x2345 
+        A.bin; a1.bin:0, a2.bin:0x8000,a3.bin:0x10000,a4.bin:0x18000 
+        B.bin; b1.bin:0, b2.bin:0x8000;                                 \x0d:0x1234, \xfa:0x2345 
 EOF
 
     exit 0
 }
 
-
 function log()
 {
-    echo ----  $@  ----
+    echo -e "\\t$@"
 }
 
 function fill_file()
@@ -35,7 +34,6 @@ function fill_file()
         return $FAIL
     fi
 
-    
     # fill 0xff
     local cnt_file=`wc --bytes $dst_file | awk '{print $1}'`
     if [ $offset -lt $cnt_file ]; then
@@ -45,29 +43,25 @@ function fill_file()
     local cnt_pad_0xff=$(($offset - $cnt_file))
 
     if [ "$cnt_pad_0xff" -gt 0 ]; then
-        log "fill 0xff to $dst_file"
-        dd if=/dev/zero bs="$cnt_pad_0xff" count=1 | tr "\000" "\377" >>"$dst_file"
+        log "fill 0xff [$cnt_pad_0xff bytes] to $dst_file"
+        dd  2>/dev/null  if=/dev/zero bs="$cnt_pad_0xff" count=1 | tr "\000" "\377" >>"$dst_file"
     fi
 
-
     # fill file
-    dd if="$src_file" >>"$dst_file"
-
+    log "fill $src_file [$(wc --bytes $src_file | awk '{print $1}') bytes] to $dst_file"
+    dd  2>/dev/null  if="$src_file" >>"$dst_file"
 
     # fill 0x00
     local cnt_file=`wc --bytes $dst_file | awk '{print $1}'`
     local cnt_pad_0x00=$((1024 - $cnt_file % 1024))
     if [ "$cnt_pad_0x00" -lt 1024 ]; then
-        log "fill 0x00 to $dst_file"
-        dd if=/dev/zero bs=$cnt_pad_0x00 count=1 >>"$dst_file"
+        log "fill 0x00 [$cnt_pad_0x00 bytes] to $dst_file"
+        dd  2>/dev/null  if=/dev/zero bs=$cnt_pad_0x00 count=1 >>"$dst_file"
     fi
 }
 
 function replace_char()
 {
-    printf "\\$1"
-    return 0
-
     local new_char="$1"
     local offset=$(($2))
     local file="$3"
@@ -77,7 +71,8 @@ function replace_char()
         return $FAIL
     fi
 
-    printf "\\$1" | dd of="$file" bs=1 seek=$offset count=1 conv=notrunc
+    log "modify $file offset[$2] --> \\\\$new_char"
+    printf "\\$1" | dd  2>/dev/null  of="$file" bs=1 seek=$offset count=1 conv=notrunc
 }
 
 # main
@@ -95,21 +90,24 @@ fi
 
 
 OLD_IFS="$IFS"
-while read new_file file_offset modify_offset
+while IFS=';' read new_file file_offset modify_offset
 do
+    if [[ "$new_file" == "#"* ]]; then continue; fi
     if [ -z "$new_file" ]; then continue; fi
 
     > "$new_file"
+    log "\\n\\n\\t\\tGenerate New File: $new_file" 
+    log "--------------------------------------------------------------"
+    
 
-    IFS=','
+    IFS=${OLD_IFS}','
     array=($file_offset)
     for item in ${array[@]}
     do
         IFS=':'
         sub_array=($item)
-        log ${sub_array[0]} ${sub_array[1]}
 
-        #fill_file  "${sub_array[0]}" "${sub_array[1]}" "$new_file" 
+        fill_file  "${sub_array[0]}" "${sub_array[1]}" "$new_file" 
 
         if [ $? -eq $FAIL ]; then
             rm -f "$new_file"
@@ -117,13 +115,12 @@ do
         fi
     done
 
-    IFS=','
+    IFS=${OLD_IFS}','
     array=($modify_offset)
     for item in ${array[@]}
     do
         IFS=':'
         sub_array=($item)
-        log ${sub_array[0]} ${sub_array[1]}
 
         replace_char  "${sub_array[0]}" "${sub_array[1]}" "$new_file" 
 
@@ -133,5 +130,6 @@ do
         fi
     done
 
-    IFS="$OLD_IFS"
 done < $conf_file
+
+log
